% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_boot.R
\name{bootstats}
\alias{bootstats}
\title{Computes Bootstrap confidence intervals}
\usage{
bootstats(b, conf = conf, type = type)
}
\arguments{
\item{b}{An object of class 'boot', containing the output
of a bootstrap calculation.}

\item{conf}{A scalar or vector containing the confidence
level(s) of the required interval(s).}

\item{type}{A vector of character strings representing
the type of intervals required. The value should be any subset
of the values c('norm','basic', 'perc', 'bca').}
}
\value{
An object of class boot
}
\description{
Computes Bootstrap confidence intervals
}
