% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{bratley1988_Fun}
\alias{bratley1988_Fun}
\title{Bratley and Fox (1988) function}
\usage{
bratley1988_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Bratley1988;textual}{sensobol} function,
which requires \emph{n} model inputs.
}
\examples{
A <- sobol_matrices(n = 100, k = 6)
Y <- bratley1988_Fun(A)
}
