% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{bratley1992_Fun}
\alias{bratley1992_Fun}
\title{Bratley, Fox and Niederreiter (1992) function}
\usage{
bratley1992_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Bratley1992;textual}{sensobol} function,
#' which requires \emph{n} model inputs.
}
\examples{
A <- sobol_matrices(n = 100, k = 4)
Y <- bratley1992_Fun(A)
}
\references{
\insertAllCited{}
}
