% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{bratley_Fun}
\alias{bratley_Fun}
\title{Bratley et al. (1992) function}
\usage{
bratley_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Bratley1992;textual}{sensobol} function.
}
\examples{
A <- sobol_matrices(n = 100, k = 4)
Y <- bratley_Fun(A)
}
\references{
\insertAllCited{}
}
