% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_boot.R
\name{create.vectors}
\alias{create.vectors}
\title{Creates two vectors with the sensitivity indices and the model
inputs.}
\usage{
create.vectors(params, second = FALSE, third = FALSE)
}
\arguments{
\item{params}{Vector with the name of the model inputs.}

\item{second}{Boolean. If second == TRUE, it computes the confidence
intervals for second-order indices. Default is second == FALSE.}

\item{third}{Boolean. If third == TRUE, it computes the confidence
intervals for second-order indices. Default is third == FALSE.}
}
\value{
A data table.
}
\description{
Creates two vectors with the sensitivity indices and the model
inputs.
}
\examples{
n <- 5000; k <- 8; R <- 10
A <- sobol_matrices(n = n, k = k, second = TRUE, third = TRUE)
Y <- sobol.Fun(A)
sens <- sobol_indices(Y = Y, params = colnames(data.frame(A)),
R = R, n = n, parallel = "no", ncpus = 1,
second = TRUE, third = TRUE)
sens.ci <- sobol_ci(sens, params = colnames(data.frame(A)),
type = "norm", conf = 0.95, second = TRUE, third = TRUE) \%>\%
cbind(create.vectors(paste("X", 1:8, sep = ""), second = TRUE, third = TRUE))

}
