% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_boot.R
\name{create_vectors}
\alias{create_vectors}
\title{Creates two vectors with the sensitivity indices and the model
inputs.}
\usage{
create_vectors(params, second = FALSE, third = FALSE)
}
\arguments{
\item{params}{Vector with the name of the model inputs.}

\item{second}{Boolean. If second == TRUE, it computes the confidence
intervals for second-order indices. Default is second == FALSE.}

\item{third}{Boolean. If third == TRUE, it computes the confidence
intervals for second-order indices. Default is third == FALSE.}
}
\value{
A data table.
}
\description{
Creates two vectors with the sensitivity indices and the model
inputs.
}
