% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{ishigami}
\alias{ishigami}
\title{Ishigami function}
\usage{
ishigami(X1, X2, X3)
}
\arguments{
\item{X1}{First model input.}

\item{X2}{Second model input.}

\item{X3}{Third model input.}
}
\value{
A numeric vector with the model output.
}
\description{
Ishigami function
}
