% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{ishigami_Mapply}
\alias{ishigami_Mapply}
\title{Ishigami function}
\usage{
ishigami_Mapply(X)
}
\arguments{
\item{X}{A data frame, data table or matrix with the three model inputs
required to run the Ishigami function.}
}
\value{
A numeric vector with the model output.
}
\description{
Ishigami function
}
\examples{
A <- sobol_matrices(n = 100, k = 3)
Y <- ishigami_Mapply(A)
}
