% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{metafunction}
\alias{metafunction}
\title{Becker (2020) metafunction}
\usage{
metafunction(data, k_2 = 0.5, k_3 = 0.2, epsilon = NULL)
}
\arguments{
\item{data}{A matrix where each column is a model input.}

\item{k_2}{Numeric, fraction of active pairwise interactions. The number
has to be between (0, 1). Default is \code{k_2 = 0.5}.}

\item{k_3}{Numeric, fraction of active threewise interactions. The number
has to be between (0, 1). Default is \code{k_2 = 0.2}.}

\item{epsilon}{Integer. It fixes the seed for the random number generator.
The default is \code{epsilon = NULL}.}
}
\value{
A vector.
}
\description{
Metafunction that randomly combines 10 univariate functions in a
multivariate function of dimension \eqn{k}. It also allows to randomize the
fraction of active pairwise and three-wise interactions.
}
\examples{
# Define settings (number of model inputs = 86)
N <- 100; params <- paste("X", 1:86, sep = "")
# Create sample matrix
mat <- sobol_matrices(N = N, params = params)
# Compute metafunction
Y <- metafunction(mat)
}
