% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{oakley_Fun}
\alias{oakley_Fun}
\title{Oakley & O'Hagan (2004) function}
\usage{
oakley_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Oakley2004;textual}{sensobol} function.
The function needs 15 model inputs. The transformation of the
distribution of the model inputs (from \eqn{U(0, 1)} to\eqn{N(0, 1)}) is
conducted internally.
}
\examples{
# Define settings
N <- 100; params <- paste("X", 1:15, sep = "")
# Create sample matrix
mat <- sobol_matrices(N = N, params = params)
# Compute Oakley and O'Hagan (2004) function
Y <- oakley_Fun(mat)
}
\references{
\insertAllCited{}
}
