% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sobol.R
\name{plot_sobol}
\alias{plot_sobol}
\title{Plot Sobol' indices.}
\usage{
plot_sobol(data, dummy = NULL, type = 1)
}
\arguments{
\item{data}{The output of \code{sobol_indices}.}

\item{dummy}{The output of the \code{sobol_dummy} function. If supplied and
\code{type = 1}, the plot includes two horizontal dashed lines showing the
highest confidence interval for the first and total-order effects of the
dummy parameter.}

\item{type}{An integer. If \code{type = 1}, it plots first and total effects.
If \code{type = 2}, it plots second-order effects. If \code{type = 3}, it plots
third-order effects. Default is \code{type = 1}.}
}
\value{
A ggplot object.
}
\description{
Plot Sobol' indices.
}
\examples{
# Define settings
N <- 1000; params <- paste("X", 1:3, sep = ""); R <- 10

# Create sample matrix
mat <- sobol_matrices(N = N, params = params)

# Compute Ishigami function
Y <- ishigami_Fun(mat)

# Compute and bootstrap Sobol' indices
ind <- sobol_indices(Y = Y, N = N, params = params, boot = TRUE, R = R)

# Plot Sobol' indices
plot_sobol(data = ind)
}
