% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{print.sensobol}
\alias{print.sensobol}
\title{Display the results obtained with the \code{sobol_indices} function.}
\usage{
\method{print}{sensobol}(x, ...)
}
\arguments{
\item{x}{A \code{sensobol} object produced by \code{sobol_indices}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function \code{print.sensobol} informs on the first and total-order
estimators used in the computations, the total number of model runs and
the sum of first-order index. It also plots the estimated results.
}
\description{
Display the results obtained with the \code{sobol_indices} function.
}
