% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to.R
\name{print.vars}
\alias{print.vars}
\title{Display the results obtained with the \code{vars_to} function.}
\usage{
\method{print}{vars}(x, ...)
}
\arguments{
\item{x}{A \code{vars} object produced by \code{vars_to}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function \code{print.vars} informs on the number of star centers,
the value of h used and the total number of model runs.. It also plots
the VARS-TO indices.
}
\description{
Display the results obtained with the \code{vars_to} function.
}
