% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_matrices.R
\name{scrambled_second}
\alias{scrambled_second}
\title{Creation of Sobol' matrix for the computation of third-order effects. It is
an internal function of "sobol_matrices".}
\usage{
scrambled_second(A, B)
}
\arguments{
\item{A}{First k Sobol' matrix.}

\item{B}{Second k Sobol' matrix.}
}
\value{
A matrix
}
\description{
Creation of Sobol' matrix for the computation of third-order effects. It is
an internal function of "sobol_matrices".
}
