% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_matrices.R
\name{scrambled_sobol}
\alias{scrambled_sobol}
\title{Creation of Sobol' matrices for the computation of
first and total order effects. It is an internal function of
"sobol_matrices".}
\usage{
scrambled_sobol(A, B)
}
\arguments{
\item{A}{The first k Sobol' matrix.}

\item{B}{The second k Sobol' matrix.}
}
\value{
A matrix.
}
\description{
Creation of Sobol' matrices for the computation of
first and total order effects. It is an internal function of
"sobol_matrices".
}
