% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensobol-package.R
\docType{package}
\name{sensobol-package}
\alias{sensobol-package}
\alias{sensobol}
\title{sensobol: Computation of Variance-Based Sensitivity Indices}
\description{
It allows to rapidly compute, bootstrap and plot up to third-order Sobol'-based sensitivity
indices using several state-of-the-art first and total-order estimators. Sobol' indices
can be computed either for models that yield a scalar as a model output or for systems of
differential equations. The package also provides a suit of benchmark tests functions
and several options to obtain publication-ready figures of the model output uncertainty
and sensitivity-related analysis.
}
\details{
A comprehensive empirical study of several total-order estimators included in sensobol can be found in \insertCite{Puyz;textual}{sensobol}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnaldpuy/sensobol}
  \item Report bugs at \url{https://github.com/arnaldpuy/sensobol/issues}
}

}
\author{
Arnald Puy (\email{arnald.puy@pm.me})

\strong{Maintainer}: Arnald Puy (\email{arnald.puy@pm.me})
}
\keyword{modeling}
\keyword{sensitivity}
\keyword{uncertainty}
