% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_boot.R
\name{sobol_ci}
\alias{sobol_ci}
\title{Computes bootstrap confidence intervals.}
\usage{
sobol_ci(b, params, type, conf, second = FALSE, third = FALSE)
}
\arguments{
\item{b}{A boot object with the computed Sobol' indices.}

\item{params}{A vector with the name of the model inputs.}

\item{type}{A vector of character strings representing
the type of intervals required. The value should be any subset
of the values c('norm','basic', 'perc', 'bca'). For more information,
check the function boot::boot.ci.}

\item{conf}{A scalar or vector containing the confidence
level(s) of the required interval(s).}

\item{second}{Boolean. If second == TRUE, it computes the confidence
intervals for second-order indices. Default is second == FALSE.}

\item{third}{Boolean. If third == TRUE, it computes the confidence
intervals for second-order indices. Default is third == FALSE.}
}
\value{
A data table.
}
\description{
Computes bootstrap confidence intervals.
}
\examples{
n <- 5000; k <- 8; R <- 10
A <- sobol_matrices(n = n, k = k, second = TRUE, third = TRUE)
Y <- sobol.Fun(A)
sens <- sobol_indices(Y = Y, params = colnames(data.frame(A)),
R = R, n = n, parallel = "no", ncpus = 1,
second = TRUE, third = TRUE)
sens.ci <- sobol_ci(sens, params = colnames(data.frame(A)),
type = "norm", conf = 0.95, second = TRUE, third = TRUE)
}
