% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_boot.R
\name{sobol_ci_temp}
\alias{sobol_ci_temp}
\title{Computes bootstrap confidence intervals}
\usage{
sobol_ci_temp(b, params, type, conf, second = FALSE, third = FALSE)
}
\arguments{
\item{b}{A boot object with the computed Sobol' indices.}

\item{params}{A vector with the name of the model inputs.}

\item{type}{A vector of character strings representing
the type of intervals required. The value should be any subset
of the values c('norm','basic', 'perc', 'bca'). For more information,
check the function boot::boot.ci.}

\item{conf}{A scalar or vector containing the confidence
level(s) of the required interval(s).}

\item{second}{Boolean. If second == TRUE, it computes the confidence
intervals for second-order indices. Default is second == FALSE.}

\item{third}{Boolean. If third == TRUE, it computes the confidence
intervals for second-order indices. Default is third == FALSE.}
}
\value{
A data table.
}
\description{
Computes bootstrap confidence intervals
}
