% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_compute_thirdS}
\alias{sobol_compute_thirdS}
\title{Computation of third-order Sobol' indices using the
Saltelli et al. 2010 estimators.}
\usage{
sobol_compute_thirdS(Y_A, Y_B, Y_ABi, Y_ABj, Y_ABk, Y_ABij, Y_ABik, Y_ABjk,
  Y_ABijk)
}
\arguments{
\item{Y_A}{Numeric vector, model output of the A matrix.}

\item{Y_B}{Numeric vector, model output of the B matrix.}

\item{Y_ABi}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the i-th,
which comes from the B matrix.}

\item{Y_ABj}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the j-th,
which comes from the B matrix.}

\item{Y_ABk}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the k-th,
which comes from the B matrix.}

\item{Y_ABij}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the i-th
and the j-th, which come from the B matrix.}

\item{Y_ABik}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the i-th
and the k-th, which come from the B matrix.}

\item{Y_ABjk}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the j-th
and the k-th, which come from the B matrix.}

\item{Y_ABijk}{Numeric vector, model output of the AB matrix,
where all the columns come from the A matrix except the i-th,
the j-th and the k-th, which come from the B matrix.}
}
\value{
The third-order indices.
}
\description{
Computation of third-order Sobol' indices using the
Saltelli et al. 2010 estimators.
}
