% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_dummy_Mapply}
\alias{sobol_dummy_Mapply}
\title{Computation of Sobol' indices for a dummy parameter.}
\usage{
sobol_dummy_Mapply(d, i)
}
\arguments{
\item{d}{Data table with the model output of
the A, B and AB matrix as columns}

\item{i}{Indices to allow bootstrap.}
}
\value{
The first and total Sobol' indices for a dummy parameter.
}
\description{
Computation of Sobol' indices for a dummy parameter.
}
