% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_dummyT}
\alias{sobol_dummyT}
\title{Computation of Sobol' indices for a dummy parameter.}
\usage{
sobol_dummyT(Y_A, Y_B, Y_AB)
}
\arguments{
\item{Y_A}{Numeric vector, model output of the A matrix.}

\item{Y_B}{Numeric vector, model output of the B matrix.}

\item{Y_AB}{Numeric vector, model output of the AB matrix,
where the AB matrix is composed of all columns of the
A matrix except the i-th, which comes from the B matrix.}
}
\value{
A numeric vector with the first and total Sobol' indices.
}
\description{
Computation of Sobol' indices for a dummy parameter.
}
