% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{sobol_Fun}
\alias{sobol_Fun}
\title{Sobol' G function}
\usage{
sobol_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Sobol1998;textual}{sensobol} function. The
function works with 8 model inputs.
}
\examples{
# Define settings
N <- 100; params <- paste("X", 1:8, sep = "")

# Create sample matrix
mat <- sobol_matrices(N = N, params = params)

# Compute Sobol' G
Y <- sobol_Fun(mat)
}
\references{
\insertAllCited{}
}
