% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_indices}
\alias{sobol_indices}
\title{Computation of first, second, third and total-effects Sobol' indices.}
\usage{
sobol_indices(Y, params, type = "jansen", R, n, parallel = "no",
  ncpus = 1, second = FALSE, third = FALSE)
}
\arguments{
\item{Y}{Numeric vector, model output.}

\item{params}{Vector with the name of the model inputs.}

\item{type}{Estimator to use. type == "saltelli" to use
the Saltelli et al. 2010 estimator, or type == "jansen"
to use the Jansen 1999 estimator. Default is type == "jansen.}

\item{R}{Integer, number of bootstrap replicas.}

\item{n}{Integer, sample size of the sample matrix.}

\item{parallel}{The type of parallel operation to be used (if any).
If missing, the default is taken from the option "boot.parallel"
(and if that is not set, "no"). For more information, check the
"parallel" option in the boot::boot function.}

\item{ncpus}{Integer: number of processes to be used in parallel operation:
typically one would chose this to the number of available CPUs.
For more information, check the "parallel" option in
the boot::boot function.}

\item{second}{Boolean. if second == "TRUE", it computes
second-order Sobol' indices.}

\item{third}{Boolean. if third == "TRUE", it computes
third-order Sobol' indices.}
}
\value{
A data.table object.
}
\description{
It computes and bootstraps up to third-order Sobol' indices
using either the Saltelli et al. 2010 or the Jansen 1999
estimator. The computation of third-order indices requires
the computation of second-order indices first. The function
relies on parallel computing.
}
\examples{
n <- 100; k <- 8; R <- 10
A <- sobol_matrices(n = n, k = k, second = TRUE, third = TRUE)
Y <- sobol.Fun(A)
sens <- sobol_indices(Y = Y, params = colnames(data.frame(A)),
R = R, n = n, parallel = "no", ncpus = 1,
second = TRUE, third = TRUE)
}
