% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_MapplyS}
\alias{sobol_MapplyS}
\title{Wrapper to allow bootstrap of the sobol_computeS function}
\usage{
sobol_MapplyS(d, i)
}
\arguments{
\item{d}{Data table with the model output of
the A, B and AB matrix as columns}

\item{i}{Indices to allow bootstrap.}
}
\value{
The first and total Sobol' indices.
}
\description{
Wrapper to allow bootstrap of the sobol_computeS function
}
