% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_ode.R
\name{sobol_ode}
\alias{sobol_ode}
\title{ODE solver in a Monte Carlo setting}
\usage{
sobol_ode(data, times, dt, state, func)
}
\arguments{
\item{data}{A sample matrix created with the \code{\link{sobol_matrices}} function.}

\item{times}{Time points at which we desire the model output.}

\item{dt}{Time increase.}

\item{state}{Value of the state variables.}

\item{func}{Function as required by the \code{\link[deSolve]{ode}} function.}
}
\value{
A matrix
}
\description{
ODE solver in a Monte Carlo setting
}
\examples{
# Define settings
 N <- 2 ^ 3 # Sample size of sample matrix
params <- c("r_b", "K", "beta", "alpha", "r_s", "K_s", "K_e", "r_e", "P", "T_e")

# Create sample matrix
dt <- sobol_matrices(N = N, params = params)

# Transform to appropriate distributions
dt[, "r_b"] <- qunif(dt[, "r_b"], 1.52, 1.6)
dt[, "K"] <- qunif(dt[, "K"], 100, 355)
dt[, "beta"] <- qunif(dt[, "beta"], 20000, 43200)
dt[, "alpha"] <- qunif(dt[, "alpha"], 1, 2)
dt[, "r_s"] <- qunif(dt[, "r_s"], 0.095, 0.15)
dt[, "K_s"] <- qunif(dt[, "K_s"], 24000, 25440)
dt[, "K_e"] <- qunif(dt[, "K_e"], 1, 1.2)
dt[, "r_e"] <- qunif(dt[, "r_e"], 0.92, 1)
dt[, "P"] <- qunif(dt[, "P"], 0.0015, 0.00195)
dt[, "T_e"] <- qunif(dt[, "T_e"], 0.7, 0.9)
times <- c(1, seq(25, 50, 25))

# Run function
\donttest{sobol_ode(data = dt, dt = 1, state = state, func = budworm_diff)}
}
