% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_second_MapplyS}
\alias{sobol_second_MapplyS}
\title{Wrapper to allow bootstrap of the sobol_compute_secondS function}
\usage{
sobol_second_MapplyS(d, i)
}
\arguments{
\item{d}{Data table with the model output of
the A, B and AB matrix as columns}

\item{i}{Indices to allow bootstrap.}
}
\value{
The second-order indices.
}
\description{
Wrapper to allow bootstrap of the sobol_compute_secondS function
}
