% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_indices.R
\name{sobol_third_MapplyJ}
\alias{sobol_third_MapplyJ}
\title{Wrapper to allow bootstrap of the sobol_compute_thirdJ function}
\usage{
sobol_third_MapplyJ(d, i)
}
\arguments{
\item{d}{Data table with the model output of
the A, B and AB matrix as columns}

\item{i}{Indices to allow bootstrap.}
}
\value{
The third-order indices.
}
\description{
Wrapper to allow bootstrap of the sobol_compute_thirdJ function
}
