% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to.R
\name{vars_to}
\alias{vars_to}
\title{Computation of VARS Total order index (VARS-TO)}
\usage{
vars_to(Y, star.centers, params, h, method = "all.step")
}
\arguments{
\item{Y}{Numeric vector, model output.}

\item{star.centers}{Number of star centers.}

\item{params}{Character vector with the name of the model inputs.}

\item{h}{Distance between pairs.}

\item{method}{Type of computation.}
}
\value{
A data table with the VARS-TO indices of each parameter.
}
\description{
It computes VARS-TO following \insertCite{Razavi2016a;textual}{sensobol} and \insertCite{Razavi2016b;textual}{sensobol}.
}
\examples{
# Define settings
star.centers <- 10; params <- paste("X", 1:3, sep = ""); h <- 0.1

# Create STAR-VARS
mat <- vars_matrices(star.centers = star.centers, params = params, h = h)

# Run model
y <- sensobol::ishigami_Fun(mat)

# Compute VARS-TO
ind <- vars_to(Y = y, star.centers = star.centers, params = params, h = h)
}
\references{
\insertAllCited{}
}
