% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{ishigami_Mapply}
\alias{ishigami_Mapply}
\title{Ishigami function}
\usage{
ishigami_Mapply(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Ishigami1990;textual}{sensobol} function,
which requires 3 model inputs. The transformation of the
distribution of the model inputs (from U(0, 1) to U(-pi, +pi)) is conducted
internally.
}
\examples{
A <- sobol_matrices(n = 100, k = 3)
Y <- ishigami_Mapply(A)
}
\references{
\insertAllCited{}
}
