% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{oakley_Fun}
\alias{oakley_Fun}
\title{Oakley & O'Hagan (2004) function}
\usage{
oakley_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Oakley2004;textual}{sensobol} function.
The function needs 15 model inputs. The transformation of the
distribution of the model inputs (from uniform to normally distributed
with mean = 0 and sd = 1) is conducted internally.
}
\examples{
A <- sobol_matrices(n = 100, k = 15)
Y <- oakley_Fun(A)
}
\references{
\insertAllCited{}
}
