% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sobol.R
\name{plot_multiscatter}
\alias{plot_multiscatter}
\title{Scatterplot matrix of model inputs and model output.}
\usage{
plot_multiscatter(data, N, Y, params, smpl = NULL)
}
\arguments{
\item{data}{A matrix, data frame or data.table, the output of \code{\link{sobol_matrices}}.}

\item{N}{The sample size of the base sample matrix used in \code{\link{sobol_matrices}}.}

\item{Y}{A numeric vector with the model output.}

\item{params}{A character vector with the name of the parameters.}

\item{smpl}{The number of simulations to plot (needed only to avoid overplotting).
The default is NULL.}
}
\value{
A ggplot2 object
}
\description{
Scatterplot matrix of model inputs and model output.
}
\examples{
# Define settings
N <- 1000; params <- paste("X", 1:3, sep = ""); R <- 10

# Create sample matrix
mat <- sobol_matrices(N = N, params = params)

# Compute Ishigami function
Y <- ishigami_Fun(mat)

# Plot scatterplot matrix
plot_multiscatter(data = mat, N = N, Y = Y, params = params)
}
