% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sobol.R
\name{plot_sobol}
\alias{plot_sobol}
\title{Plot Sobol' first and total-order indices}
\usage{
plot_sobol(x, type = 1)
}
\arguments{
\item{x}{A data.table}

\item{type}{An integer. If type == 1, it plots first and total effects.
If type == 2, it plots second-order effects. If type == 3, it plots
third-order effects. Default is type == 1.}
}
\value{
A ggplot object
}
\description{
Plot Sobol' first and total-order indices
}
\examples{
n <- 5000; k <- 8; R <- 10
A <- sobol_matrices(n = n, k = k, second = TRUE, third = TRUE)
Y <- sobol.Fun(A)
sens <- sobol_indices(Y = Y, params = colnames(data.frame(A)),
R = R, n = n, parallel = "no", ncpus = 1,
second = TRUE, third = TRUE)
sens.ci <- sobol_ci(sens, params = colnames(data.frame(A)),
type = "norm", conf = 0.95, second = TRUE, third = TRUE) \%>\%
cbind(create.vectors(paste("X", 1:8, sep = ""), second = TRUE, third = TRUE))
plot_sobol(sens.ci, type = 2)
}
