% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sobol.R
\name{plot_uncertainty}
\alias{plot_uncertainty}
\title{Plot model output uncertainty}
\usage{
plot_uncertainty(Y, n = NULL)
}
\arguments{
\item{Y}{A numeric vector with the model output.}

\item{n}{An integer with the initial sample size.}
}
\value{
a ggplot2 object.
}
\description{
It creates an histogram with the model output distribution.
}
\examples{
# Define settings:
}
