% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sobol.R
\name{plot_uncertainty}
\alias{plot_uncertainty}
\title{Plot model output uncertainty}
\usage{
plot_uncertainty(Y)
}
\arguments{
\item{Y}{A numeric vector with the model output}
}
\value{
a ggplot2 object
}
\description{
Plot model output uncertainty
}
\examples{
n <- 500; k <- 8
A <- sobol_matrices(n = n, k = k)
Y <- sobol.Fun(A)
plot_uncertainty(Y)
}
