% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{sobol_Fun}
\alias{sobol_Fun}
\title{Sobol' G function}
\usage{
sobol_Fun(X)
}
\arguments{
\item{X}{A data frame or numeric matrix.}
}
\value{
A numeric vector with the model output.
}
\description{
It implements the \insertCite{Sobol1998;textual}{sensobol} function. In this
case, the function works with 8 model inputs.
}
\examples{
A <- sobol_matrices(n = 100, k = 8)
Y <- sobol_Fun(A)
}
\references{
\insertAllCited{}
}
