% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_matrices.R
\name{sobol_matrices}
\alias{sobol_matrices}
\title{Creation of the sample matrices}
\usage{
sobol_matrices(
  matrices = c("A", "B", "AB"),
  N,
  params,
  order = "first",
  cluster = NULL
)
}
\arguments{
\item{matrices}{Vector with the required matrices. The default
is \code{matrices = c("A", "B", "AB")}.}

\item{N}{Integer, initial ample size of the base Sobol' matrix.}

\item{params}{Vector with the name of the model inputs.}

\item{order}{One of "first", "second" or "third" to create a matrix to
compute first, second or up to third-order Sobol' indices. The default is
\code{order = "first"}.}

\item{cluster}{List of vectors including the model inputs that form part of the
cluster/s. The default is \code{cluster = NULL}}
}
\value{
A matrix.
}
\description{
It creates the sample matrices to compute Sobol' first and total-order indices.
If needed, it also creates the sample matrices required to compute second and
third-order indices. It uses \insertCite{Sobol1967;textual}{sensobol} quasi-random number sequences.
}
\details{
The function generates a \eqn{(N, 2k)} matrix using Sobol' quasi-random
number sequences, for \eqn{i=1,2,...,k} parameters. The first \emph{k}-matrix is
the \strong{A} matrix and the remaining \emph{k}-matrix, the \strong{B}
matrix. If \code{matrices} includes an "AB" ("BA") matrix and \code{order = "first"},
the function also generates \emph{k} \strong{A}_B^{j} (\strong{B}_A^{j})
matrices, where all collumns come from \strong{A} (\strong{B}) except the
\emph{j}-th, which comes from \strong{B} (\strong{A}).

If \code{matrices} includes an "AB" ("BA") matrix and \code{order = "second"}
or \code{order = "third"}, the function will also create \eqn{n} extra
matrices, where \eqn{n = k! / 2!(k-2)!} or \eqn{n = k! /3!(k-3)!} respectively.
These are needed to compute second or third-order Sobol' indices.

The argument \code{matrices} can include either a single element
(i.e. \code{matrices = "A"}, " \code{matrices = "B"},  \code{matrices = "AB"}
or \code{matrices = "BA"}), or a vector with a combination of matrices;
for instance, \code{matrices = c("A", "AB")}, \code{matrices = c("A", "AB", "BA")},
etc. The selection of the estimator to compute first and total-order indices
will define which matrices are required:
\itemize{
\item If the estimator of choice in \code{\link{sobol_indices}} is the Azzini estimator,
\code{matrices = c("A", "B", "AB", "BA")}.
\item If the estimator in \code{\link{sobol_indices}} is the \insertCite{Janon2014;textual}{sensobol},
\code{matrices = c("A", "AB", "BA")}.
\item If the estimator in \code{\link{sobol_indices}} for first-order indices is either the
\insertCite{Saltelli2010a;textual}{sensobol} or the \insertCite{Jansen1999;textual}{sensobol},
and the estimator for total order indices is either the
\insertCite{Homma1996;textual}{sensobol}, the
\insertCite{Jansen1999;textual}{sensobol} or the
\insertCite{Sobol2001;textual}{sensobol}, \code{matrices = c("A", "B", "AB")}.
}
}
\examples{
# Define settings
N <- 100; params <- paste("X", 1:10, sep = ""); order <-  "third"

# Create sample matrix
mat <- sobol_matrices(N = N, params = params, order = order)
}
\references{
\insertAllCited{}
}
\seealso{
Check the function \code{\link{sobol}} in the package \code{randtoolbox}
to see how the Sobol' quasi-random number sequences are constructed.
}
