% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_matrices.R
\name{sobol_matrices}
\alias{sobol_matrices}
\title{Creation of the sample matrices}
\usage{
sobol_matrices(
  matrices = c("A", "B", "AB"),
  N,
  params,
  order = "first",
  type = "QRN",
  ...
)
}
\arguments{
\item{matrices}{Vector with the required matrices. The default
is \code{matrices = c("A", "B", "AB")}.}

\item{N}{Integer, initial ample size of the base sample matrix.}

\item{params}{Character vector with the name of the model inputs.}

\item{order}{One of "first", "second" or "third" to create a matrix to
compute first, second or up to third-order Sobol indices. The default is
\code{order = "first"}.}

\item{type}{Approach to construct the sample matrix. Options are:
\itemize{
\item \code{type = "QRN"}: It uses \insertCite{Sobol1967;textual}{sensobol} Quasi-Random Numbers.
through a call to the function \code{\link{sobol}} of the \code{randtoolbox} package.
\item \code{type = "LHS"}: It uses a Latin Hypercube Sampling Design
\insertCite{McKay1979}{sensobol} through a call
to the function \code{\link{randomLHS}} of the \code{lhs} package.
\item \code{type = "R"}: It uses random numbers.
}}

\item{...}{Further arguments in \code{\link{sobol}}.}
}
\value{
A matrix where each column is a model input and each row a sampling point.
}
\description{
It creates the sample matrices to compute Sobol' first and total-order indices.
If needed, it also creates the sample matrices required to compute second and
third-order indices.
}
\details{
Before calling \code{sobol_matrices}, the user must decide which estimators
will be used to compute first and total-order indices, for this option conditions
the design of the sample matrix and therefore the argument \code{matrices}.
See the vignette for further details on the specific sampling designs required by
the estimators.

The function can select one of the following sampling designs:
\itemize{
\item \eqn{\mathbf{A}}, \eqn{\mathbf{B}}, \eqn{\mathbf{A}_B^{(i)}}.
\item \eqn{\mathbf{A}}, \eqn{\mathbf{B}}, \eqn{\mathbf{B}_A^{(i)}}.
\item \eqn{\mathbf{A}}, \eqn{\mathbf{B}}, \eqn{\mathbf{A}_B^{(i)}}, \eqn{\mathbf{B}_A^{(i)}}.
}

The function initially creates an \eqn{(N, 2k)} matrix according to the approach defined by
\code{type}, where the leftmost and the rightmost \eqn{k} columns are respectively allocated
to the \eqn{\mathbf{A}} and the \eqn{\mathbf{B}} matrix. Depending on the sampling design, it
also creates \eqn{k} \eqn{\mathbf{A}_B^{(i)}} (\eqn{\mathbf{B}_A^{(i)}}) matrices, where all
columns come from \eqn{\mathbf{A}} (\eqn{\mathbf{B}}) except the \eqn{i}-th, which comes from
\eqn{\mathbf{B}} (\eqn{\mathbf{A}}).
}
\examples{
# Define settings
N <- 100; params <- paste("X", 1:10, sep = ""); order <-  "third"

# Create sample matrix
mat <- sobol_matrices(N = N, params = params, order = order)
}
\references{
\insertAllCited{}
}
