% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_matrices.R
\name{sobol_matrices}
\alias{sobol_matrices}
\title{Creation of the sample matrices}
\usage{
sobol_matrices(
  matrices = c("A", "B", "AB"),
  N,
  params,
  order = "first",
  type = "QRN",
  ...
)
}
\arguments{
\item{matrices}{Vector with the required matrices. The default
is \code{matrices = c("A", "B", "AB")}.}

\item{N}{Integer, initial ample size of the base sample matrix.}

\item{params}{Character vector with the name of the model inputs.}

\item{order}{One of "first", "second" or "third" to create a matrix to
compute first, second or up to third-order Sobol indices. The default is
\code{order = "first"}.}

\item{type}{Approach to construct the sample matrix. Options are:
\itemize{
\item \code{type = "QRN"}: It uses Sobol' Quasi-Random Numbers \insertCite{Sobol1967}{sensobol}
through a call to the function \code{\link{sobol}} of the \code{randtoolbox} package.
\item \code{type = "LHS"}: It uses a Latin Hypercube Sampling Design
\insertCite{McKay1979}{sensobol} through a call
to the function \code{\link{randomLHS}} of the \code{lhs} package.
\item \code{type = "R"}: It uses random numbers.
}}

\item{...}{Further arguments in \code{\link{sobol}}.}
}
\value{
A matrix where each column is a model input and each row a sampling point.
}
\description{
It creates the sample matrices to compute Sobol' first and total-order indices.
If needed, it also creates the sample matrices required to compute second and
third-order indices.
}
\details{
Before calling \code{sobol_matrices}, the user must decide which estimators
will be used to compute first and total-order indices, for this option conditions
the design of the sample matrix and therefore the argument \code{matrices}.
See the vignette for further details on the specific sampling designs required by
the estimators.
}
\examples{
# Define settings
N <- 100; params <- paste("X", 1:10, sep = ""); order <-  "third"

# Create sample matrix
mat <- sobol_matrices(N = N, params = params, order = order)
}
\references{
\insertAllCited{}
}
