% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_matrices.R
\name{sobol_matrices}
\alias{sobol_matrices}
\title{Creation of Sobol' matrices to compute first, second, third and total-
order effects.}
\usage{
sobol_matrices(n, k, second = FALSE, third = FALSE)
}
\arguments{
\item{n}{Integer, sample size of the Sobol' matrix.}

\item{k}{Integer, number of model inputs.}

\item{second}{Boolean. If second == TRUE, it creates the scrambled
matrix required to compute second-order effects. Default is second == FALSE.}

\item{third}{Boolean. If third == TRUE, it creates the scrambled
matrix required to compute third-order effects. Default is second == FALSE.}
}
\value{
A matrix
}
\description{
Creation of Sobol' matrices to compute first, second, third and total-
order effects.
}
\examples{
sobol_matrices(n = 1000, k = 8, second = TRUE, third = TRUE)
}
