% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_ode.R
\name{sobol_ode}
\alias{sobol_ode}
\title{Solver for ordinary differential equations}
\usage{
sobol_ode(params, times, state, func, ...)
}
\arguments{
\item{params}{Vector with the name of the model inputs.}

\item{times}{Vector with the time sequences at which the model output is wanted.}

\item{state}{Initial values of the state variables.}

\item{func}{An R function as defined by \code{ode}.}

\item{...}{Additional arguments passed to \code{ode}.}
}
\value{
A matrix with the output values.
}
\description{
Solver for ordinary differential equations
}
\examples{
# Define the model: the Lotka-Volterra system of equations
lotka_volterra_fun <- function(t, state, parameters) {
with(as.list(c(state, parameters)), {
  dX <- r * X * (1 - X / K) - alpha * X * Y
  dY <- -m * Y + theta * X * Y
  list(c(dX, dY))
 })
 }

# Define the settings of the sensitivity analysis
N <- 2 ^ 5 # Sample size of sample matrix
params <- c("r", "alpha", "m", "theta", "K", "X", "Y") # Parameters
R <- 100 # Number of bootstrap replicas

# Construct the sample matrix
mat <- sobol_matrices(N = N, params = params)

# Define the timesteps
times <- seq(5, 15, 5)

# Run the model
}
